requirements.txt
blinker==1.4
certifi==2021.5.30
charset-normalizer==2.0.6
docutils==0.17.1
feedgenerator==1.9.2
idna==3.2
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
pelican==4.6.0
Pygments==2.10.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
Unidecode==1.3.2
urllib3==1.26.7
